modernizacion de gradle y agp, UI con feedback

+++ app/src/main/java/com/cluster/facelabs/clusterface/FaceHandler.java
@@ -3,20 +3,21 @@ package com.cluster.facelabs.clusterface;
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.net.Uri;
-import android.os.Environment;
-import android.support.annotation.NonNull;
 import android.util.Log;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import com.bumptech.glide.Glide;
-import com.bumptech.glide.request.target.SimpleTarget;
+import com.bumptech.glide.request.target.CustomTarget;
 import com.bumptech.glide.request.transition.Transition;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.ml.vision.FirebaseVision;
-import com.google.firebase.ml.vision.common.FirebaseVisionImage;
-import com.google.firebase.ml.vision.face.FirebaseVisionFace;
-import com.google.firebase.ml.vision.face.FirebaseVisionFaceDetector;
-import com.google.firebase.ml.vision.face.FirebaseVisionFaceDetectorOptions;
+import com.google.mlkit.vision.common.InputImage;
+import com.google.mlkit.vision.face.Face;
+import com.google.mlkit.vision.face.FaceDetection;
+import com.google.mlkit.vision.face.FaceDetector;
+import com.google.mlkit.vision.face.FaceDetectorOptions;
 
 import org.apache.commons.io.FilenameUtils;
 
@@ -32,99 +33,84 @@ public class FaceHandler {
 
     int mIdx = -1;
     File[] files;
-    /*[TODO]
-     * Queue up-to 5 instead of just 1
-     * */
     int mQueueCounter = 0;
     int mQueueMax = 5;
 
     private Context mContext;
-    FirebaseVisionFaceDetector mFaceDetector;
+    FaceDetector mFaceDetector;
 
     public FaceHandler(Context context){
         mContext = context;
 
-        /**face detector options*/
-        float minFaceSize = 0.05f*(1+minFaceSizeSeekbar.getProgress());
+        float minFaceSize = 0.05f * (1 + minFaceSizeSeekbar.getProgress());
         int performanceMode;
         if(faceDetectModeSwitch.isEnabled())
-            performanceMode = FirebaseVisionFaceDetectorOptions.FAST;
+            performanceMode = FaceDetectorOptions.PERFORMANCE_MODE_FAST;
         else
-            performanceMode = FirebaseVisionFaceDetectorOptions.ACCURATE;
+            performanceMode = FaceDetectorOptions.PERFORMANCE_MODE_ACCURATE;
 
-        FirebaseVisionFaceDetectorOptions options =
-                new FirebaseVisionFaceDetectorOptions.Builder()
+        FaceDetectorOptions options =
+                new FaceDetectorOptions.Builder()
                         .setPerformanceMode(performanceMode)
                         .setMinFaceSize(minFaceSize)
                         .build();
 
-        /**get face detector*/
-        mFaceDetector = FirebaseVision.getInstance().getVisionFaceDetector(options);
+        mFaceDetector = FaceDetection.getClient(options);
     }
 
-    /**get an instance of FirebaseVisionImage from imagepath or bitmap*/
-    private FirebaseVisionImage getFirebaseVisionImage(Uri imagePath, Bitmap bitmap){
-        FirebaseVisionImage image = null;
+    private InputImage getFirebaseVisionImage(Uri imagePath, Bitmap bitmap){
+        InputImage image = null;
         if(imagePath != null){
             try {
-                image = FirebaseVisionImage.fromFilePath(mContext, imagePath);
+                image = InputImage.fromFilePath(mContext, imagePath);
             } catch (IOException e) {
-                e.printStackTrace();
-                Utils.showToast(mContext, "Unable to create FireBase Image from uri!");
+                Utils.showToast(mContext, "No se pudo crear la imagen de entrada desde la ruta especificada: " + e.toString());
             }
         }else{
-            image = FirebaseVisionImage.fromBitmap(bitmap);
+            image = InputImage.fromBitmap(bitmap, 0);
         }
         return image;
     }
 
-    /**get faces from an image*/
     private void runFaceRecognition(final Uri imagePath,
                                     final Bitmap bitmap,
                                     final String imageName){
-        FirebaseVisionImage image = getFirebaseVisionImage(imagePath, bitmap);
-        /**start detection with callbacks*/
-        mFaceDetector.detectInImage(image)
+        InputImage image = getFirebaseVisionImage(imagePath, bitmap);
+        if (image == null) {
+            next();
+            return;
+        }
+        mFaceDetector.process(image)
                 .addOnSuccessListener(
-                        new OnSuccessListener<List<FirebaseVisionFace>>() {
+                        new OnSuccessListener<List<Face>>() {
                             @Override
-                            public void onSuccess(List<FirebaseVisionFace> firebaseVisionFaces) {
-                                processFaceRecognitionResult(firebaseVisionFaces,
-                                        imagePath, bitmap, imageName);
-
+                            public void onSuccess(List<Face> faces) {
+                                processFaceRecognitionResult(faces, imagePath, bitmap, imageName);
                                 next();
-
                             }
                         })
                 .addOnFailureListener(
                         new OnFailureListener() {
                             @Override
                             public void onFailure(@NonNull Exception e) {
-                                // Task failed with an exception
-                                e.printStackTrace();
-                                Utils.showToast(mContext, "Face detector failed!");
+                                Utils.showToast(mContext, "Detección de rostros fallida en el procesamiento: " + e.toString());
                                 next();
                             }
                         });
     }
 
-    /**process the output of "runFaceRecognition"*/
-    private void processFaceRecognitionResult(List<FirebaseVisionFace> faces,
+    private void processFaceRecognitionResult(List<Face> faces,
                                               Uri imagePath, Bitmap bitmap, String imageName){
         if(faces.size() == 0){
-            Log.d("finding faces", "No faces found!");
             return;
         }
 
-        /**get the image from which crops will be extracted
-         * based on identified bounding boxes*/
         Bitmap inputImg;
         if(imagePath != null) {
             try {
                 inputImg = Utils.getBitmapFromUri(mContext, imagePath);
             } catch (IOException e) {
-                e.printStackTrace();
-                Utils.showToast(mContext, "Face : Cannot load image from uri!");
+                Utils.showToast(mContext, "Error al cargar bitmap original: " + e.toString());
                 return;
             }
         }else{
@@ -132,32 +118,34 @@ public class FaceHandler {
         }
 
         int i = 0;
-        for(FirebaseVisionFace face : faces){
-            /**get bounding box details*/
+        for(Face face : faces){
             int top = face.getBoundingBox().top;
             int left = face.getBoundingBox().left;
             int width = face.getBoundingBox().width();
             int height = face.getBoundingBox().height();
 
-            /**create a bitmap for identified face*/
+            if (top < 0) top = 0;
+            if (left < 0) left = 0;
+            if (left + width > inputImg.getWidth()) width = inputImg.getWidth() - left;
+            if (top + height > inputImg.getHeight()) height = inputImg.getHeight() - top;
+
             try {
                 Bitmap croppedFace = Bitmap.createBitmap(inputImg, left, top, width, height);
                 Utils.saveImage(croppedFace, imageName, String.valueOf(i++));
             }catch (IllegalArgumentException e){
-                Log.d("finding", "Illegal argument to crop!");
+                Log.e("FaceHandler", "Error al procesar el recorte de coordenadas: " + e.toString(), e);
             }
         }
 
         MainActivity.faceProgressbar.incrementProgressBy(1);
     }
 
-    /**find faces for all images in the input directory*/
     public void getCrops(){
         String inputDirPath = Utils.getInputPath();
 
         File inputDir = new File(inputDirPath);
         if(!inputDir.exists()) {
-            Utils.showToast(mContext, "ERROR : Input folder not found!");
+            Utils.showToast(mContext, "Directorio de entrada no localizado: " + inputDirPath);
             return;
         }
 
@@ -168,27 +156,30 @@ public class FaceHandler {
             }
         });
         if(files == null){
-            Utils.showToast(mContext, "ERROR : No files found in the input folder!");
+            Utils.showToast(mContext, "No se encontraron archivos en la carpeta de entrada.");
             return;
         }
 
-        /**remove the crops from the crops folder
-         * for which the input image is no longer present*/
         File cropsDir = new File(Utils.getCropsPath());
         if(cropsDir.exists())
         {
-            for(File crop : cropsDir.listFiles(new FilenameFilter() {
+            File[] existingCrops = cropsDir.listFiles(new FilenameFilter() {
                 @Override
                 public boolean accept(File dir, String name) {
                     return !name.equals(".nomedia");
                 }
-            }))
-            {
-                String cropName = FilenameUtils.removeExtension(crop.getName());
-                String possibleInputName  = cropName.substring(0, cropName.lastIndexOf('_')) + "." + FilenameUtils.getExtension(crop.getName());
-                File possibleInput = new File(inputDirPath + "/" + possibleInputName);
-                if(!possibleInput.exists())
-                    crop.delete();
+            });
+            if (existingCrops != null) {
+                for(File crop : existingCrops)
+                {
+                    String cropName = FilenameUtils.removeExtension(crop.getName());
+                    if (cropName.contains("_")) {
+                        String possibleInputName  = cropName.substring(0, cropName.lastIndexOf('_')) + "." + FilenameUtils.getExtension(crop.getName());
+                        File possibleInput = new File(inputDirPath + "/" + possibleInputName);
+                        if(!possibleInput.exists())
+                            crop.delete();
+                    }
+                }
             }
         }
 
@@ -202,31 +193,47 @@ public class FaceHandler {
 
     private void next(){
         mIdx++;
-        if(mIdx >= files.length) return;
+        if(mIdx >= files.length) {
+            new android.os.Handler(android.os.Looper.getMainLooper()).post(new Runnable() {
+                @Override
+                public void run() {
+                    MainActivity.faceFeedbackText.setText("Detección finalizada de " + files.length + " fotos.");
+                }
+            });
+            return;
+        }
         final String fileName = FilenameUtils.removeExtension(files[mIdx].getName());
-        /**if crops for this input image have been already found, skip*/
+        final String currentName = files[mIdx].getName();
+        final int currentCount = mIdx + 1;
+        final int totalCount = files.length;
+        final int percentage = (currentCount * 100) / totalCount;
+
+        new android.os.Handler(android.os.Looper.getMainLooper()).post(new Runnable() {
+            @Override
+            public void run() {
+                MainActivity.faceFeedbackText.setText("Procesando: " + currentName + "\n" + currentCount + " / " + totalCount + " (" + percentage + "%)");
+            }
+        });
+
         String cropCheckName = Utils.getCropsPath() + "/" + fileName + "_0.jpg";
         File cropCheck = new File(cropCheckName);
-        Log.e("finding faces", cropCheckName);
         if(cropCheck.exists()){
             MainActivity.faceQueueProgressbar.incrementProgressBy(1);
             MainActivity.faceProgressbar.incrementProgressBy(1);
-            Log.d("finding faces", "Crop exists!");
             next();
         }else{
-            Log.d("finding faces", "Processing...");
-
-            /**runFaceRecognition(Uri.fromFile(files[i]), null);
-             * this gives error and loads images with the wrong orientation etc
-             * cropping seems to always fix this
-             * using glide to load bitmap from file with center cropping*/
             Glide.with(mContext).asBitmap().load(files[mIdx])
-                    .into(new SimpleTarget<Bitmap>() {
+                    .into(new CustomTarget<Bitmap>() {
                         @Override
-                        public void onResourceReady(Bitmap resource, Transition<? super Bitmap> transition) {
-                            runFaceRecognition(null, resource, fileName);}});
+                        public void onResourceReady(@NonNull Bitmap resource, @Nullable Transition<? super Bitmap> transition) {
+                            runFaceRecognition(null, resource, fileName);
+                        }
+
+                        @Override
+                        public void onLoadCleared(@Nullable android.graphics.drawable.Drawable placeholder) {
+                        }
+                    });
             MainActivity.faceQueueProgressbar.incrementProgressBy(1);
         }
     }
-
 }
